backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
sqlalchemy>=2.0.36
google-generativeai==0.3.1
python-docx==1.1.0
python-pptx==0.6.23
pydantic>=2.9.2,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
python-dotenv==1.0.0
email-validator>=2.2.0
psycopg2-binary>=2.9.0

